+++ mod.ts
@@ -1,2 +1,5 @@
 export * from "./fetch_wrapper.ts";
-export type { ExtendedRequestInit } from "./extended_request_init.ts";
+export type {
+  ExtendedRequest,
+  ExtendedRequestInit,
+} from "./extended_request_init.ts";